Extract targetMeshGeometry into own file
